Honour XDG user download directory
Depending on the language used on the user’s machine, the default
download directory might not be ~/downloads, which is itself
non-standard.

The script should first look up the XDG_DOWNLOAD_DIR variable which
might hold the name of the user’s default directory --- for instance,
French-speaking users of Linux distributions such as Ubuntu have
~/Téléchargements as their default download directory.

If such a directory does not exist, then the script should fall back
to the default download directory in Linux distributions, ~/Downloads

This commit implements such a behaviour in scitopdf.

It also adds a file-local variable for Emacs users to enforce the use
of tabs instead of spaces in indentation.